refactor:Tidy up scalar module (#255)
# Rationale for this change

MontScalar and Scalar logic is currently disorganized. We should consolidate this logic.

# What changes are included in this PR?

Moved the Scalar trait to its own "scalar" file within base/scalar.
