+++ crates/proof-of-sql/src/base/mod.rs
@@ -10,6 +10,7 @@ pub mod math;
 pub(crate) mod polynomial;
 pub(crate) mod proof;
 pub(crate) mod ref_into;
+/// This module contains the `Scalar` trait as well as the main, generic, implementations of it.
 pub mod scalar;
 mod serialize;
 pub(crate) use serialize::{impl_serde_for_ark_serde_checked, impl_serde_for_ark_serde_unchecked};
